ebpf: Always use `release` profile
Using `dev` profile, even after tuning flags (to make it identical
with `release`) still causes problems and is confusing for people.

And even when we have BTF support in future, it'd be better and
less confusing to set `debug=2` in `release` profile (because we
would want debug info, but still keep the same opt-level and other
options).