Use `target_arch = "bpf"` for `panic_handler`-cfg
The `panic_handler` defined in the eBPF crate is only required if we are
compiling for the bpf target. When compiling for the host-system, be it to
run tests or by compiling the entire workspace with `cargo check --workspace`
we don't want to define the `panic_handler`, otherwise it will clash with
the one defined by the standard library.

Related: https://github.com/aya-rs/book/issues/214