chore: Add LICENSE
This adds a LICENSE to the generated code, but also to the code included
in this repository. To make things simple, userspace code is Apache/MIT
licensed - like Aya. eBPF code is MIT/GPL which is GPL compatible to
allow use of all the eBPF helper functions, but is also permissively
licensed via the MIT license.

Signed-off-by: Dave Tucker <dave@dtucker.co.uk>